+++ backend/src/voogle/routers/__init__.py
@@ -8,4 +8,5 @@ from . import app as app
 from . import assistant as assistant
 from . import local as local
 from . import media as media
+from . import resource as resource
 from . import users as users